customer_feedback_analyzer/scraper.py
import requests
from bs4 import BeautifulSoup
import time

def scrape_yelp(url, max_pages=3):
    """Scrape reviews from a Yelp business page."""
    reviews = []
    page = 0
    headers = {'User-Agent': 'Mozilla/5.0'}  # Avoid bot detection

    while page < max_pages:
        page_url = f"{url}?start={page * 20}"  # Yelp uses ?start= for pagination
        try:
            response = requests.get(page_url, headers=headers)
            response.raise_for_status()
            soup = BeautifulSoup(response.text, 'html.parser')

            # Extract business name (example selector, may need adjustment)
            business_name = soup.find('h1', class_='css-1se8maq').text.strip()

            # Find review elements (adjust selectors based on current Yelp HTML)
            review_elements = soup.find_all('div', class_='review__09f24__oHr9V')
            if not review_elements:
                break

            for elem in review_elements:
                review_text_elem = elem.find('p', class_='comment__09f24__gu0rG')
                rating_elem = elem.find('div', class_='i-stars__09f24__M1AR7')
                date_elem = elem.find('span', class_='css-chan6m')

                review_text = review_text_elem.text.strip() if review_text_elem else 'N/A'
                rating = float(rating_elem['aria-label'].split()[0]) if rating_elem else 0
                review_date = date_elem.text.strip() if date_elem else 'N/A'

                reviews.append({
                    'business_name': business_name,
                    'platform': 'yelp',
                    'review_text': review_text,
                    'rating': rating,
                    'review_date': review_date
                })

            # Check for next page
            next_link = soup.find('a', class_='next-link')
            if not next_link:
                break

            page += 1
            time.sleep(1)  # Polite delay to avoid overloading server

        except Exception as e:
            print(f"Error scraping {page_url}: {e}")
            break

    return reviews

def scrape_google(url):
    """Placeholder for Google Reviews scraping."""
    # API Placeholder:
    # Instead of scraping, use the Google Places API for reliable data
    # Example:
    # api_key = 'YOUR_API_KEY'
    # place_id = extract_place_id(url)
    # reviews = fetch_reviews_from_google_api(place_id, api_key)
    # return reviews

    print("Google Reviews scraping not fully implemented due to dynamic content.")
    print("Consider using the Google Places API with an API key for production use.")
    # Basic attempt with BeautifulSoup (limited functionality)
    headers = {'User-Agent': 'Mozilla/5.0'}
    response = requests.get(url, headers=headers)
    soup = BeautifulSoup(response.text, 'html.parser')
    # Note: Google Reviews load dynamically; this won’t capture all reviews
    # Adjust selectors as needed, but Selenium or API is recommended
    return []  # Placeholder return