customer_feedback_analyzer/requirements.txt
beautifulsoup4
requests
pandas
matplotlib
wordcloud
textblob
python-dateutil